% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/martix_converter.R
\name{matrix_conversion}
\alias{matrix_conversion}
\title{Conversione matrice}
\usage{
matrix_conversion(input_file, output_file)
}
\arguments{
\item{input_file}{File in entrata}

\item{output_file}{File in uscita}
}
\description{
Conversione matrice
}
